refactor: Add dto packages
