+++ src/main/java/roomescape/config/WebConfig.java
@@ -3,7 +3,6 @@ package roomescape.config;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.member.MemberRequest;
 
 import java.util.List;
