+++ src/main/java/roomescape/member/dto/LoginMember.java
@@ -1,6 +1,8 @@
-package roomescape.member;
+package roomescape.member.dto;
 
-public class LoginMember extends Member{
+import roomescape.member.Member;
+
+public class LoginMember extends Member {
     public LoginMember(Long id, String name, String email, String password) {
         super(id, name, email, password);
     }
