+++ src/main/java/roomescape/reservation/ReservationRepository.java
@@ -2,7 +2,6 @@ package roomescape.reservation;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import roomescape.member.LoginMember;
 
 import java.util.List;
